Make breakpoints resolver reusable and improve error handling
