+++ packages/material-ui-lab/src/MasonryItem/MasonryItem.test.js
@@ -60,7 +60,7 @@ describe('<MasonryItem />', () => {
           paddingBottom: Number(theme.spacing(3).replace('px', '')) - 1,
         },
         width: '100%',
-        [`& *`]: {
+        [`& > *`]: {
           objectFit: 'cover',
           width: '100%',
           height: '100%',
