+++ packages/material-ui-system/src/index.js
@@ -2,7 +2,7 @@ export { css, keyframes, GlobalStyles, StyledEngineProvider } from '@material-ui
 export { default as borders } from './borders';
 export * from './borders';
 export { default as breakpoints } from './breakpoints';
-export { handleBreakpoints, mergeBreakpointsInOrder } from './breakpoints';
+export { handleBreakpoints, mergeBreakpointsInOrder, resolveBreakpointValues } from './breakpoints';
 export { default as compose } from './compose';
 export { default as display } from './display';
 export { default as flexbox } from './flexbox';
